Remove DrawableComponent and associated functionalities
Several changes were made in the system, mostly focused on removing the DrawableComponent and all related functionalities. The systems and classes using it were either modified or completely removed. ButtonRenderer and various components were moved to a new "Renderers" subdirectory to better organize the code structure. This change is part of an effort to simplify the codebase, and thereby improve maintainability and readability.