aider: Add IsConfigurableComponent to entities created with TransformComponent in EntityBuilderExtensions.
# Aider chat conversation:

USER: Update all the entities who have a TransformationComponent to also have an ConfigurableComponent

ASSISTANT: To update all entities that have a `TransformComponent` to also have an `IsConfigurableComponent`, we need to modify the part of the code where entities are created or managed. Since the `EntityBuilderExtensions` class seems to be responsible for creating and configuring entities, I will add the `IsConfigurableComponent` to entities within the `WithTransform` method, which is likely called when entities with a `TransformComponent` are created.

Here is the diff for the `EntityBuilderExtensions` class:

```diff
--- Nazar/Factories/EntityBuilder.cs
+++ Nazar/Factories/EntityBuilder.cs
@@ ... @@
 public static class EntityBuilderExtensions
 {
     public static Entity WithTransform(this Entity entity, Pose? position = null)
     {
+        entity.Set<IsConfigurableComponent>();
         if (position.HasValue)
         {
             entity.Set(new TransformComponent(position.Value));
         }
         return entity;
     }
     // ... other methods ...
 }
```

This change ensures that any entity that is given a `TransformComponent` through the `WithTransform` method will also be tagged with the `IsConfigurableComponent`. If entities are created or managed in a different part of the codebase, please provide the relevant file or instructions on where to implement the changes.